Add abstract autoencoder class #27. Rename FlexibleAutoencoder to FeedforwardAutoencoder
