+++ clustpy/deep/autoencoders/__init__.py
@@ -1,11 +1,11 @@
 from .stacked_autoencoder import StackedAutoencoder
-from .flexible_autoencoder import FlexibleAutoencoder
+from .feedforward_autoencoder import FeedforwardAutoencoder
 from .variational_autoencoder import VariationalAutoencoder
 from .neighbor_encoder import NeighborEncoder
 from .convolutional_autoencoder import ConvolutionalAutoencoder
 
 
-__all__ = ['FlexibleAutoencoder',
+__all__ = ['FeedforwardAutoencoder',
            'StackedAutoencoder',
            'VariationalAutoencoder',
            'ConvolutionalAutoencoder',
