+++ clustpy/deep/autoencoders/_abstract_autoencoder.py
@@ -24,7 +24,7 @@ class FullyConnectedBlock(torch.nn.Module):
     activation_fn : torch.nn.Module
         activation function from torch.nn, set the activation function for the hidden layers, if None then it will be linear (default: None)
     bias : bool
-        set False if you do not want to use a bias term in the linear layers (default: None)
+        set False if you do not want to use a bias term in the linear layers (default: True)
     output_fn : torch.nn.Module
         activation function from torch.nn, set the activation function for the last layer, if None then it will be linear (default: None)
 
@@ -79,28 +79,12 @@ class FullyConnectedBlock(torch.nn.Module):
         return forwarded
 
 
-class FlexibleAutoencoder(torch.nn.Module):
+class _AbstractAutoencoder(torch.nn.Module):
     """
-    A flexible feedforward autoencoder.
+    An abstract autoencoder class that can be used by other autoencoder implementations.
 
     Parameters
     ----------
-    layers : list
-        list of the different layer sizes from input to embedding, e.g. an example architecture for MNIST [784, 512, 256, 10], where 784 is the input dimension and 10 the embedding dimension.
-        If decoder_layers are not specified then the decoder is symmetric and goes in the same order from embedding to input.
-    batch_norm : bool
-        Set True if you want to use torch.nn.BatchNorm1d (default: False)
-    dropout : float
-        Set the amount of dropout you want to use (default: None)
-    activation_fn : torch.nn.Module
-        activation function from torch.nn, set the activation function for the hidden layers, if None then it will be linear (default: torch.nn.LeakyReLU)
-    bias : bool
-        set False if you do not want to use a bias term in the linear layers (default: True)
-    decoder_layers : list
-        list of different layer sizes from embedding to output of the decoder. If set to None, will be symmetric to layers (default: None)
-    decoder_output_fn : torch.nn.Module
-        activation function from torch.nn, set the activation function for the decoder output layer, if None then it will be linear.
-        E.g. set to torch.nn.Sigmoid if you want to scale the decoder output between 0 and 1 (default: None)
     reusable : bool
         If set to true, deep clustering algorithms will optimize a copy of the autoencoder and not the autoencoder itself.
         Ensures that the same autoencoder can be used by multiple deep clustering algorithms.
@@ -108,78 +92,48 @@ class FlexibleAutoencoder(torch.nn.Module):
 
     Attributes
     ----------
-    encoder : FullyConnectedBlock
-        encoder part of the autoencoder, responsible for embedding data points (class is FullyConnectedBlock)
-    decoder : FullyConnectedBlock
-        decoder part of the autoencoder, responsible for reconstructing data points from the embedding (class is FullyConnectedBlock)
     fitted  : bool
         indicates whether the autoencoder is already fitted
     reusable : bool
-        indicates whether the autoencoder should be reused by mutliple deep clustering algorithms
-
-    References
-    ----------
-    E.g. Ballard, Dana H. "Modular learning in neural networks." Aaai. Vol. 647. 1987.
+        indicates whether the autoencoder should be reused by multiple deep clustering algorithms
     """
 
-    def __init__(self, layers: list, batch_norm: bool = False, dropout: float = None,
-                 activation_fn: torch.nn.Module = torch.nn.LeakyReLU, bias: bool = True, decoder_layers: list = None,
-                 decoder_output_fn: torch.nn.Module = None, reusable: bool = True):
-        super(FlexibleAutoencoder, self).__init__()
+    def __init__(self, reusable: bool = True):
+        super(_AbstractAutoencoder, self).__init__()
         self.fitted = False
         self.reusable = reusable
-        if decoder_layers is None:
-            decoder_layers = layers[::-1]
-        if (layers[-1] != decoder_layers[0]):
-            raise ValueError(
-                f"Innermost hidden layer and first decoder layer do not match, they are {layers[-1]} and {decoder_layers[0]} respectively.")
-        if (layers[0] != decoder_layers[-1]):
-            raise ValueError(
-                f"Output and input dimension do not match, they are {layers[0]} and {decoder_layers[-1]} respectively.")
-        # Initialize encoder
-        self.encoder = FullyConnectedBlock(layers=layers, batch_norm=batch_norm, dropout=dropout,
-                                           activation_fn=activation_fn, bias=bias, output_fn=None)
-
-        # Inverts the list of layers to make symmetric version of the encoder
-        self.decoder = FullyConnectedBlock(layers=decoder_layers, batch_norm=batch_norm, dropout=dropout,
-                                           activation_fn=activation_fn, bias=bias,
-                                           output_fn=decoder_output_fn)
 
     def encode(self, x: torch.Tensor) -> torch.Tensor:
         """
-        Apply the encoder function to x.
+        Placeholder for an encode function of an autoencoder.
 
         Parameters
         ----------
         x : torch.Tensor
             input data point, can also be a mini-batch of points
-        
+
         Returns
         -------
-        embedded : torch.Tensor
-            the embedded data point with dimensionality embedding_size
+        x : torch.Tensor
+            should return the embedded data point
         """
-        assert x.shape[1] == self.encoder.layers[0], "Input layer of the encoder does not match input sample"
-        embedded = self.encoder(x)
-        return embedded
+        return x
 
     def decode(self, embedded: torch.Tensor) -> torch.Tensor:
         """
-        Apply the decoder function to embedded.
+        Placeholder for a decode function of an autoencoder.
 
         Parameters
         ----------
         embedded : torch.Tensor
             embedded data point, can also be a mini-batch of embedded points
-        
+
         Returns
         -------
-        decoded : torch.Tensor
-            returns the reconstruction of embedded
+        embedded : torch.Tensor
+            should return the reconstruction of embedded
         """
-        assert embedded.shape[1] == self.decoder.layers[0], "Input layer of the decoder does not match input sample"
-        decoded = self.decoder(embedded)
-        return decoded
+        return embedded
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -190,7 +144,7 @@ class FlexibleAutoencoder(torch.nn.Module):
         ----------
         x : torch.Tensor
             input data point, can also be a mini-batch of embedded points
-        
+
         Returns
         -------
         reconstruction : torch.Tensor
@@ -207,7 +161,7 @@ class FlexibleAutoencoder(torch.nn.Module):
 
         Parameters
         ----------
-        batch: list
+        batch : list
             the different parts of a dataloader (id, samples, ...)
         loss_fn : torch.nn.modules.loss._Loss
             loss function to be used for reconstruction
@@ -232,7 +186,7 @@ class FlexibleAutoencoder(torch.nn.Module):
                  device: torch.device = torch.device("cpu")) -> torch.Tensor:
         """
         Evaluates the autoencoder.
-        
+
         Parameters
         ----------
         dataloader : torch.utils.data.DataLoader
@@ -241,7 +195,7 @@ class FlexibleAutoencoder(torch.nn.Module):
             loss function to be used for reconstruction
         device : torch.device
             device to be trained on (default: torch.device('cpu'))
-        
+
         Returns
         -------
         loss: torch.Tensor
@@ -260,12 +214,12 @@ class FlexibleAutoencoder(torch.nn.Module):
             data_eval: np.ndarray = None, dataloader: torch.utils.data.DataLoader = None,
             evalloader: torch.utils.data.DataLoader = None, optimizer_class: torch.optim.Optimizer = torch.optim.Adam,
             loss_fn: torch.nn.modules.loss._Loss = torch.nn.MSELoss(), patience: int = 5,
-            scheduler: torch.optim.lr_scheduler = None, scheduler_params: dict = None,
+            scheduler: torch.optim.lr_scheduler = None, scheduler_params: dict = {},
             device: torch.device = torch.device("cpu"), model_path: str = None,
-            print_step: int = 0) -> 'FlexibleAutoencoder':
+            print_step: int = 0) -> '_AbstractAutoencoder':
         """
         Trains the autoencoder in place.
-        
+
         Parameters
         ----------
         n_epochs : int
@@ -292,7 +246,7 @@ class FlexibleAutoencoder(torch.nn.Module):
             learning rate scheduler that should be used.
             If torch.optim.lr_scheduler.ReduceLROnPlateau is used then the behaviour is matched by providing the validation_loss calculated based on samples from evalloader (default: None)
         scheduler_params : dict
-            dictionary of the parameters of the scheduler object (default: None)
+            dictionary of the parameters of the scheduler object (default: {})
         device : torch.device
             device to be trained on (default: torch.device('cpu'))
         model_path : str
@@ -302,8 +256,8 @@ class FlexibleAutoencoder(torch.nn.Module):
 
         Returns
         -------
-        self : FlexibleAutoencoder
-            this instance of the FlexibleAutoencoder
+        self : _AbstractAutoencoder
+            this instance of the autoencoder
 
         Raises
         ----------
