+++ clustpy/deep/autoencoders/tests/test_feedforward_autoencoder.py
@@ -1,19 +1,15 @@
-from clustpy.deep.autoencoders import FlexibleAutoencoder
+from clustpy.deep.autoencoders import FeedforwardAutoencoder
 from clustpy.data import create_subspace_data
 import torch
 
 
-def test_simple_flexible_autoencoder():
+def test_simple_feedforward_autoencoder():
     data, _ = create_subspace_data(1500, subspace_features=(3, 50), random_state=1)
-    embedding_dim = 10
-    # Test fitting
-    autoencoder = FlexibleAutoencoder(layers=[data.shape[1], 128, 64, embedding_dim])
-    assert autoencoder.fitted is False
-    autoencoder.fit(n_epochs=5, optimizer_params={"lr": 1e-3}, data=data)
-    assert autoencoder.fitted is True
-    # Test encoding
     batch_size = 256
     data_batch = torch.Tensor(data[:batch_size])
+    embedding_dim = 10
+    autoencoder = FeedforwardAutoencoder(layers=[data.shape[1], 128, 64, embedding_dim])
+    # Test encoding
     embedded = autoencoder.encode(data_batch)
     assert embedded.shape == (batch_size, embedding_dim)
     # Test decoding
@@ -22,3 +18,7 @@ def test_simple_flexible_autoencoder():
     # Test forwarding
     forwarded = autoencoder.forward(data_batch)
     assert torch.equal(decoded, forwarded)
+    # Test fitting
+    assert autoencoder.fitted is False
+    autoencoder.fit(n_epochs=3, optimizer_params={"lr": 1e-3}, data=data)
+    assert autoencoder.fitted is True
