+++ clustpy/deep/tests/test_train_utils.py
@@ -1,4 +1,4 @@
-from clustpy.deep.autoencoders import FlexibleAutoencoder, VariationalAutoencoder
+from clustpy.deep.autoencoders import FeedforwardAutoencoder, VariationalAutoencoder
 from clustpy.deep._train_utils import get_trained_autoencoder, _get_default_layers
 from clustpy.deep.tests._helpers_for_tests import _get_test_dataloader
 from clustpy.data import create_subspace_data
@@ -27,7 +27,7 @@ def test_get_trained_autoencoder():
                                  optimizer_class=torch.optim.Adam,
                                  loss_fn=torch.nn.MSELoss(), input_dim=data.shape[1], embedding_size=10)
     # Check output of get_trained_autoencoder
-    assert type(ae) is FlexibleAutoencoder
+    assert type(ae) is FeedforwardAutoencoder
     assert ae.fitted == True
 
 
@@ -52,7 +52,7 @@ def test_get_trained_autoencoder_with_custom_ae():
     dataloader = _get_test_dataloader(data, 256, True, False)
     # Get trained version of custom AE
     device = torch.device('cpu')
-    ae = FlexibleAutoencoder(layers=[data.shape[1], 256, 128, 64, 10], reusable=False)
+    ae = FeedforwardAutoencoder(layers=[data.shape[1], 256, 128, 64, 10], reusable=False)
     assert ae.fitted == False
     encoder_0_params = ae.encoder.block[0].weight.data.detach().clone()
     decoder_0_params = ae.decoder.block[0].weight.data.detach().clone()
@@ -61,7 +61,7 @@ def test_get_trained_autoencoder_with_custom_ae():
                                  loss_fn=torch.nn.MSELoss(), input_dim=data.shape[1], embedding_size=10,
                                  autoencoder=ae)
     # Check output of get_trained_autoencoder
-    assert type(ae_out) is FlexibleAutoencoder
+    assert type(ae_out) is FeedforwardAutoencoder
     assert ae_out.fitted == True
     assert not torch.equal(encoder_0_params, ae_out.encoder.block[0].weight.data)
     assert not torch.equal(decoder_0_params, ae_out.decoder.block[0].weight.data)
@@ -74,7 +74,7 @@ def test_get_trained_autoencoder_with_custom_pretrained_ae():
     dataloader = _get_test_dataloader(data, 256, True, False)
     # Get same pretrained version out of get_trained_autoencoder
     device = torch.device('cpu')
-    ae = FlexibleAutoencoder(layers=[data.shape[1], 256, 128, 64, 10], reusable=True)
+    ae = FeedforwardAutoencoder(layers=[data.shape[1], 256, 128, 64, 10], reusable=True)
     assert ae.fitted == False
     ae.fit(dataloader=dataloader, optimizer_params={"lr":1e-3}, n_epochs=5, device=device,
            optimizer_class=torch.optim.Adam,
@@ -86,7 +86,7 @@ def test_get_trained_autoencoder_with_custom_pretrained_ae():
                                  loss_fn=torch.nn.MSELoss(), input_dim=data.shape[1], embedding_size=10,
                                  autoencoder=ae)
     # Check output of get_trained_autoencoder
-    assert type(ae_out) is FlexibleAutoencoder
+    assert type(ae_out) is FeedforwardAutoencoder
     assert ae_out.fitted == True
     assert torch.equal(encoder_0_params, ae_out.encoder.block[0].weight.data)
     assert torch.equal(decoder_0_params, ae_out.decoder.block[0].weight.data)
